uploaded 1-5 assignments
1. Write a Java program that uses polymorphism by defining an interface called Shape
with methods to calculate the area and perimeter of a shape. Then create classes
that implement the Shape interface for different types of shapes, such as circles and
triangles.
2. Write a Java program to invoke parent class constructor from a child class. Create
Child class object and parent class constructor must be invoked. Demonstrate by
writing a program. Also explain key points about Constructor.
3. Write a Java programme that takes an integer from the user and throws an exception
if it is negative.Demonstrate Exception handling of same program as solution.
4. Create a Java program that simulates a bank account. The program should allow
users to deposit and withdraw money, check their balance.
5. Demonstrate the difference between abstract class and interface by writing programs
as well as in keypoints.